refactor(mobile): outbox extraction + MQTT session split (OJD-1470)
Replace the upload-queue/worker pair with a transactional Outbox
backed by SQLite, and split the MQTT transport into a reusable Paho
session. Adds wide-event tracing and dev-seed plumbing on top.
